Updates to be consistent with repo. Rename files to kebab case. Refactor imports. Update eslint to allow double quotes if escaping can be avoided.
