+++ b2b-app/.eslintrc.js
@@ -23,7 +23,7 @@ module.exports = {
   rules: {
     indent: ['error', 2, { SwitchCase: 1 }],
     'linebreak-style': ['error', 'unix'],
-    quotes: ['error', 'single'],
+    quotes: ['error', 'single', 'avoidEscape'],
     semi: ['error', 'never'],
   },
 }
