+++ b2b-app/imports/ui/admin/register/forms/about-form.js
@@ -1,13 +1,11 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import { emergencyFormBridge } from './formSchemas'
-
 import { makeStyles } from '@material-ui/core/styles'
-import Button from '@material-ui/core/Button'
-import Paper from '@material-ui/core/Paper'
-
+import { Button, Paper } from '@material-ui/core'
 import { AutoForm, AutoFields, ErrorsField } from 'uniforms-material'
+
+import { aboutFormBridge } from './form-schemas'
 
 const useStyles = makeStyles((theme) => ({
   formButtons: {
@@ -19,12 +17,12 @@ const useStyles = makeStyles((theme) => ({
   },
 }))
 
-const EmergencyForm = ({ onSubmit, model, onBack }) => {
+const AboutForm = ({ onSubmit, model, onBack }) => {
   const classes = useStyles()
 
   return (
     <Paper style={{ padding: '2rem', paddingBottom: '0', marginBottom: '1rem' }}>
-      <AutoForm schema={emergencyFormBridge} onSubmit={onSubmit} model={model}>
+      <AutoForm schema={aboutFormBridge} onSubmit={onSubmit} model={model}>
         <AutoFields />
         <ErrorsField />
         <div className={classes.buttonGroup}>
@@ -50,10 +48,10 @@ const EmergencyForm = ({ onSubmit, model, onBack }) => {
   )
 }
 
-EmergencyForm.propTypes = {
+AboutForm.propTypes = {
   onSubmit: PropTypes.func.isRequired,
   model: PropTypes.object.isRequired,
   onBack: PropTypes.func.isRequired,
 }
 
-export { EmergencyForm }
+export { AboutForm }
