+++ b2b-app/imports/ui/admin/register/forms/index.js
@@ -1,0 +1,4 @@
+export { AboutForm } from './about-form'
+export { ConfirmForm } from './confirm-form'
+export { ContactForm } from './contact-form'
+export { EmergencyForm } from './emergency-form'
